abis/linux: add statfs and statvfs structs
