--- options/posix/include/bits/posix/fsblkcnt_t.h
@@ -1,7 +1,0 @@
-
-#ifndef MLIBC_FSBLKCNT_T_H
-#define MLIBC_FSBLKCNT_T_H
-
-typedef unsigned int fsblkcnt_t;
-
-#endif // MLIBC_FSBLKCNT_T_H
