--- options/posix/include/bits/posix/fsfilcnt_t.h
@@ -1,7 +1,0 @@
-
-#ifndef MLIBC_FSFILCNT_T_H
-#define MLIBC_FSFILCNT_T_H
-
-typedef unsigned int fsfilcnt_t;
-
-#endif // MLIBC_FSFILCNT_T_H
