+++ options/posix/include/sys/statfs.h
@@ -1,4 +1,3 @@
-
 #ifndef _SYS_STATFS_H
 #define _SYS_STATFS_H
 
@@ -6,27 +5,7 @@
 extern "C" {
 #endif
 
-#include <bits/posix/fsblkcnt_t.h>
-#include <bits/posix/fsfilcnt_t.h>
-
-typedef struct __mlibc_fsid {
-	int __val[2];
-} fsid_t;
-
-struct statfs {
-	unsigned long f_type;
-	unsigned long f_bsize;
-	fsblkcnt_t f_blocks;
-	fsblkcnt_t f_bfree;
-	fsblkcnt_t f_bavail;
-
-	fsfilcnt_t f_files;
-	fsfilcnt_t f_ffree;
-	fsid_t f_fsid;
-	unsigned long f_namelen;
-	unsigned long f_frsize;
-	unsigned long d_flags;
-};
+#include <abi-bits/statfs.h>
 
 int statfs(const char *, struct statfs *);
 int fstatfs(int, struct statfs *);
